Added pause, resume functionality and default configuration loading.
By clicking the Start menu item on the GUI when no simulation is
created, it tries to load the default configuration found in the
resources. Pause menu item will make the simulation thread wait
for notification sent when Start clicked or on exit. Also, loading
methods in the Game class will not start the simulation after creating
it. That can be done with the startSimulation() method of the Game
class.